requirements.txt
# Core dependencies
pypdf>=3.15.0
pdfminer.six>=20231228
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# LLM integrations
openai>=1.3.0
anthropic>=0.7.0
tiktoken>=0.5.0

# Data processing
pandas>=2.1.0
numpy>=1.25.0

# Database
sqlite-utils>=3.35.0

# Web framework for GUI
fastapi>=0.104.0
uvicorn>=0.24.0

# Frontend assets
jinja2>=3.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0